Bufixes, optimizations and cleaning

+++ scripts/clean.sh
@@ -1,5 +1,8 @@
-i686-w64-mingw32-strip src/openocd.exe
-mkdir output
-cp src/openocd.exe output/
-cp -r tcl/ output/scripts
+if [ -f "src/openocd.exe" ]; then
+  ${STRIP:-i686-w64-mingw32-strip} src/openocd.exe
+  mkdir -p output
+  cp src/openocd.exe output/
+  cp -r tcl/ output/scripts
+fi
+make clean
 make distclean
